Revert "xgcd -> xgcd_minimal"
This reverts commit eaf7528c6884097fbc1ddd6d25d38b77a3aefb28.
