+++ fmpz/test/t-xgcd.c
@@ -1,6 +1,5 @@
 /*
     Copyright (C) 2009 William Hart
-    Copyright (C) 2021 Albin Ahlbäck
 
     This file is part of FLINT.
 
@@ -20,333 +19,13 @@
 int
 main(void)
 {
-    int i, j, result;
-    fmpz_t maxval;
-    fmpz_t d, x, y, a, b;
-    fmpz_t tx, ty;
-
+    int i, result;
     FLINT_TEST_INIT(state);
-    fmpz_set_str(maxval, "999999999999999999999999999999999999999999", 10);
 
-    flint_printf("xgcd_minimal....");
+    flint_printf("xgcd....");
     fflush(stdout);
 
-
-    fmpz_init(d);
-    fmpz_init(x);
-    fmpz_init(y);
-    fmpz_init(a);
-    fmpz_init(b);
-
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check small (0, 0)
-                                       * gives (0, 0, 0) */
-    result = (fmpz_is_zero(d)
-           && fmpz_is_zero(x)
-           && fmpz_is_zero(y)
-           && fmpz_is_zero(a)
-           && fmpz_is_zero(b));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    _fmpz_promote_val(a);
-    _fmpz_promote_val(b);
-    fmpz_zero(a);
-    fmpz_zero(b);
-
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check promoted (0, 0)
-                                       * gives (0, 0, 0) */
-    result = (fmpz_is_zero(d)
-           && fmpz_is_zero(x)
-           && fmpz_is_zero(y));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d %d\n",fmpz_is_zero(d));
-        flint_printf("x %d\n",fmpz_is_zero(x));
-        flint_printf("y %d\n",fmpz_is_zero(y));
-        flint_printf("a %d\n",fmpz_is_zero(a));
-        flint_printf("b %d\n",fmpz_is_zero(b));
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_one(a);
-    fmpz_set_str(b, "999", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (1, b) for small b gives
-                                       * (1, **long expressions**) */
-
-    /* i = |sgn((b - 1) * (b + 1))| */
-    fmpz_init_set(tx, b);
-    fmpz_init_set(ty, b);
-    fmpz_sub_si(tx, tx, 1);
-    fmpz_add_ui(ty, ty, 1);
-    fmpz_mul(tx, tx, ty);
-    i = abs(fmpz_sgn(tx));
-
-    /* j = sgn(b) * (sgn(b + 1) - sgn(b - 1)) */
-    fmpz_set(tx, b);
-    fmpz_set(ty, b);
-    fmpz_sub_si(tx, tx, 1);
-    fmpz_add_ui(ty, ty, 1);
-    j = fmpz_sgn(b) * (fmpz_sgn(ty) - fmpz_sgn(tx));
-
-    result = (fmpz_is_one(d)
-           && fmpz_equal_si(x, i)
-           && fmpz_equal_si(y, j));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_one(a);
-    fmpz_set_str(b, "99999999999999999999999999999999999999999999", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (1, b) for big b gives
-                                       * (1, **long expressions**) */
-
-    /* i = |sgn((b - 1) * (b + 1))| */
-    fmpz_init_set(tx, b);
-    fmpz_init_set(ty, b);
-    fmpz_sub_si(tx, tx, 1);
-    fmpz_add_ui(ty, ty, 1);
-    fmpz_mul(tx, tx, ty);
-    i = abs(fmpz_sgn(tx));
-
-    /* j = sgn(b) * (sgn(b + 1) - sgn(b - 1)) */
-    fmpz_set(tx, b);
-    fmpz_set(ty, b);
-    fmpz_sub_si(tx, tx, 1);
-    fmpz_add_ui(ty, ty, 1);
-    j = fmpz_sgn(b) * (fmpz_sgn(ty) - fmpz_sgn(tx));
-
-    result = (fmpz_is_one(d)
-           && fmpz_equal_si(x, i)
-           && fmpz_equal_si(y, j));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "999", 10);
-    fmpz_one(b);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, 1) for small a gives
-                                       * (1, 0, 1) */
-
-    result = (fmpz_is_one(d)
-           && fmpz_is_zero(x)
-           && fmpz_is_one(y));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "9771293786198273698712639876198276398716298736999", 10);
-    fmpz_one(b);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, 1) for big a gives
-                                       * (1, 0, 1) */
-
-    result = (fmpz_is_one(d)
-           && fmpz_is_zero(x)
-           && fmpz_is_one(y));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "977", 10);
-    fmpz_set_str(b, "977", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, a) for small a gives
-                                       * (|a|, 0, sgn(a)) */
     
-    fmpz_abs(tx, a);
-    result = (fmpz_equal(d, tx)
-           && fmpz_is_zero(x)
-           && fmpz_equal_si(y, fmpz_sgn(a)));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "977817236128736817263876121723681726387618273681", 10);
-    fmpz_set_str(b, "977817236128736817263876121723681726387618273681", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, a) for big a gives
-                                       * (|a|, 0, sgn(a)) */
-    
-    fmpz_abs(tx, a);
-    result = (fmpz_equal(d, tx)
-           && fmpz_is_zero(x)
-           && fmpz_equal_si(y, fmpz_sgn(a)));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "97177", 10);
-    fmpz_set_str(b, "-97177", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, -a) for small a gives
-                                       * (|a|, 0, -sgn(a)) */
-    
-    fmpz_abs(tx, a);
-    result = (fmpz_equal(d, tx)
-           && fmpz_is_zero(x)
-           && fmpz_equal_si(y, -fmpz_sgn(a)));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d = "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-    fmpz_set_str(a, "9778172361736817263876121723681726387618273681", 10);
-    fmpz_set_str(b, "-9778172361736817263876121723681726387618273681", 10);
-    fmpz_xgcd_minimal(d, x, y, a, b); /* check (a, -a) for big a gives
-                                       * (|a|, 0, -sgn(a)) */
-    
-    fmpz_abs(tx, a);
-    result = (fmpz_equal(d, tx)
-           && fmpz_is_zero(x)
-           && fmpz_equal_si(y, -fmpz_sgn(a)));
-
-    if (!result)
-    {
-        flint_printf("FAIL:\n\n");
-        flint_printf("d =akjsdn "), fmpz_print(d), flint_printf("\n");
-        flint_printf("x = "), fmpz_print(x), flint_printf("\n");
-        flint_printf("y = "), fmpz_print(y), flint_printf("\n");
-        flint_printf("a = "), fmpz_print(a), flint_printf("\n");
-        flint_printf("b = "), fmpz_print(b), flint_printf("\n");
-        abort();
-    }
-
-
-
-    /* Test the switch (f, g) -> (g, f) is ok */
-    for (i = 0; i < 1000 * flint_test_multiplier(); i++)
-    {
-        fmpz_t d1, d2, a1, a2, b1, b2, f, g, tmpa, tmpb;
-
-        fmpz_init(d1);
-        fmpz_init(a1);
-        fmpz_init(b1);
-        fmpz_init(d2);
-        fmpz_init(a2);
-        fmpz_init(b2);
-        fmpz_init(f);
-        fmpz_init(g);
-        fmpz_init(tmpa);
-        fmpz_init(tmpb);
-
-        fmpz_randm(f, state, maxval);
-        fmpz_randm(g, state, maxval);
-        if (fmpz_is_zero(g)) /* ensure not dividing by 0 */
-            fmpz_one(g);
-        if (n_randint(state, 2)) fmpz_neg(g, g);
-        if (n_randint(state, 2)) fmpz_neg(f, f);
-
-        fmpz_xgcd_minimal(d1, a1, b1, f, g);
-        fmpz_xgcd_minimal(d2, b2, a2, g, f);
-
-        fmpz_divexact(tmpa, g, d1);
-        fmpz_divexact(tmpb, f, d1);
-        result = (fmpz_equal(d1, d2)
-               && (fmpz_is_zero(g) || fmpz_is_zero(f)
-               || (fmpz_cmpabs(a1, tmpa) <= 0 && fmpz_cmpabs(b1, tmpb) <= 0)
-               || (fmpz_cmpabs(a2, tmpa) <= 0 && fmpz_cmpabs(b2, tmpb) <= 0)));
-        if (!result)
-        {
-            flint_printf("FAIL:\n\n");
-            flint_printf("d1 = "), fmpz_print(d1), flint_printf("\n");
-            flint_printf("a1 = "), fmpz_print(a1), flint_printf("\n");
-            flint_printf("b1 = "), fmpz_print(b1), flint_printf("\n");
-            flint_printf("d2 = "), fmpz_print(d2), flint_printf("\n");
-            flint_printf("a2 = "), fmpz_print(a2), flint_printf("\n");
-            flint_printf("b2 = "), fmpz_print(b2), flint_printf("\n");
-            flint_printf("f = "), fmpz_print(f), flint_printf("\n");
-            flint_printf("g = "), fmpz_print(g), flint_printf("\n");
-            abort();
-        }
-
-        fmpz_clear(d1);
-        fmpz_clear(a1);
-        fmpz_clear(b1);
-        fmpz_clear(d2);
-        fmpz_clear(a2);
-        fmpz_clear(b2);
-        fmpz_clear(f);
-        fmpz_clear(g);
-        fmpz_init(tmpa);
-        fmpz_init(tmpb);
-    }
-
 
     /* Test aliasing of d and f, a and g */
     for (i = 0; i < 1000 * flint_test_multiplier(); i++)
@@ -370,8 +49,8 @@ main(void)
         fmpz_set(f, F);
         fmpz_set(g, G);
 
-        fmpz_xgcd_minimal(d, a, b, f, g);
-        fmpz_xgcd_minimal(f, g, c, f, g);
+        fmpz_xgcd(d, a, b, f, g);
+        fmpz_xgcd(f, g, c, f, g);
 
         result = (fmpz_equal(d, f)
                && fmpz_equal(b, c)
@@ -400,7 +79,6 @@ main(void)
         fmpz_clear(G);
     }
 
-
     /* Test aliasing of a and f, d and g */
     for (i = 0; i < 1000 * flint_test_multiplier(); i++)
     {
@@ -423,8 +101,8 @@ main(void)
         fmpz_set(f, F);
         fmpz_set(g, G);
 
-        fmpz_xgcd_minimal(d, a, b, f, g);
-        fmpz_xgcd_minimal(g, f, c, f, g);
+        fmpz_xgcd(d, a, b, f, g);
+        fmpz_xgcd(g, f, c, f, g);
 
         result = (fmpz_equal(d, g)
                && fmpz_equal(b, c)
@@ -453,7 +131,6 @@ main(void)
         fmpz_clear(G);
     }
 
-
     /* Test aliasing of d and f, b and g */
     for (i = 0; i < 1000 * flint_test_multiplier(); i++)
     {
@@ -476,8 +153,8 @@ main(void)
         fmpz_set(f, F);
         fmpz_set(g, G);
 
-        fmpz_xgcd_minimal(d, a, b, f, g);
-        fmpz_xgcd_minimal(f, c, g, f, g);
+        fmpz_xgcd(d, a, b, f, g);
+        fmpz_xgcd(f, c, g, f, g);
 
         result = (fmpz_equal(d, f) 
                && fmpz_equal(a, c)
@@ -506,7 +183,6 @@ main(void)
         fmpz_clear(G);
     }
 
-
     /* Test aliasing of b and f, d and g */
     for (i = 0; i < 1000 * flint_test_multiplier(); i++)
     {
@@ -529,8 +205,8 @@ main(void)
         fmpz_set(f, F);
         fmpz_set(g, G);
 
-        fmpz_xgcd_minimal(d, a, b, f, g);
-        fmpz_xgcd_minimal(g, c, f, f, g);
+        fmpz_xgcd(d, a, b, f, g);
+        fmpz_xgcd(g, c, f, f, g);
 
         result = (fmpz_equal(d, g)
                && fmpz_equal(a, c)
@@ -559,7 +235,6 @@ main(void)
         fmpz_clear(G);
     }
 
-
     /* Test a f  + b g == d and d >= 0 */
     for (i = 0; i < 1000 * flint_test_multiplier(); i++)
     {
@@ -579,7 +254,7 @@ main(void)
         if (n_randint(state, 2)) fmpz_neg(g, g);
         if (n_randint(state, 2)) fmpz_neg(f, f);
         
-        fmpz_xgcd_minimal(d, a, b, f, g);
+        fmpz_xgcd(d, a, b, f, g);
 
         fmpz_mul(t1, a, f);
         fmpz_mul(t2, b, g);
@@ -608,9 +283,9 @@ main(void)
         fmpz_clear(t2);
     }
 
-
     FLINT_TEST_CLEANUP(state);
     
     flint_printf("PASS\n");
     return 0;
 }
+
