+++ fmpz/xgcd.c
@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2021 Albin Ahlbäck
+    Copyright (C) 2012 William Hart
 
     This file is part of FLINT.
 
@@ -10,90 +10,94 @@
 */
 
 #include <gmp.h>
+#include "flint.h"
+#include "ulong_extras.h"
 #include "fmpz.h"
 
 void
 fmpz_xgcd(fmpz_t d, fmpz_t a, fmpz_t b, const fmpz_t f, const fmpz_t g)
 {
-    if (fmpz_is_zero(g))
+    fmpz_t t1, t2;
+    fmpz *f1, *g1;
+
+    if (fmpz_is_zero(f))
     {
-        slong sgn = fmpz_sgn(f);
-        fmpz_abs(d, f);
-        fmpz_set_si(a, sgn);
-        fmpz_zero(b);
-        return;
-    }
-    else if (fmpz_is_zero(f))
-    {
-        slong sgn = fmpz_sgn(g);
+        int sign = fmpz_sgn(g);
         fmpz_abs(d, g);
-        fmpz_set_si(b, sgn);
-        fmpz_zero(a);
-        return;
+        fmpz_set_ui(a, 0);
+        if (sign == 0)
+            fmpz_set_ui(b, 0);
+        else if (sign > 0)
+            fmpz_set_ui(b, 1);
+        else
+            fmpz_set_si(b, -1);
     }
-    else if (d == f || a == f || b == f || d == g || a == g || b == g)
+    else if (fmpz_cmpabs(f, g) == 0)
     {
-        fmpz_t d2, a2, b2;
-        fmpz_init(d2);
-        fmpz_init(a2);
-        fmpz_init(b2);
-        fmpz_xgcd(d2, a2, b2, f, g);
-        fmpz_swap(d, d2);
-        fmpz_swap(a, a2);
-        fmpz_swap(b, b2);
-        fmpz_clear(d2);
-        fmpz_clear(a2);
-        fmpz_clear(b2);
-        return;
+        if (fmpz_sgn(g) > 0)
+        {
+            fmpz_set(d, g);
+            fmpz_set_ui(b, 1);
+        }
+        else
+        {
+            fmpz_neg(d, g);
+            fmpz_set_si(b, -1);
+        }
+        fmpz_set_si(a, 0);
     }
-
-    _fmpz_promote(d);
-    _fmpz_promote(a);
-    _fmpz_promote(b);
-
-    if (!COEFF_IS_MPZ(*f) && !COEFF_IS_MPZ(*g))  /* both are small */
+    else
     {
-        mpz_t mf, mg;
-        ulong tf = FLINT_ABS(*f);
-        ulong tg = FLINT_ABS(*g);
+        int sign1 = fmpz_sgn(f);
+        int sign2 = fmpz_sgn(g);
 
-        mf->_mp_d = (mp_limb_t *) &tf;
-        mg->_mp_d = (mp_limb_t *) &tg;
-        mf->_mp_size  = fmpz_sgn(f);
-        mg->_mp_size  = fmpz_sgn(g);
+        fmpz_init(t1);
+        fmpz_init(t2);
 
-        mpz_gcdext(COEFF_TO_PTR(*d), COEFF_TO_PTR(*a), COEFF_TO_PTR(*b),
-                   mf, mg);
+        /* support aliasing */
+        if (d == f || a == f || sign1 < 0)
+        {
+            f1 = t1;
+            if (sign1 < 0)
+                fmpz_neg(f1, f);
+            else
+                fmpz_set(f1, f);
+        }
+        else
+            f1 = (fmpz *) f;
+
+        if (d == g || a == g || sign2 < 0)
+        {
+            g1 = t2;
+            if (sign2 < 0)
+                fmpz_neg(g1, g);
+            else
+                fmpz_set(g1, g);
+        }
+        else
+            g1 = (fmpz *) g;
+
+        if (fmpz_cmp(f1, g1) < 0)
+        {
+            fmpz_gcdinv(d, a, f1, g1);
+            fmpz_mul(t1, a, f1);
+            fmpz_sub(t1, d, t1);
+            fmpz_divexact(b, t1, g1);
+        }
+        else                    /* g < f */
+        {
+            fmpz_gcdinv(d, b, g1, f1);
+            fmpz_mul(t2, b, g1);
+            fmpz_sub(t2, d, t2);
+            fmpz_divexact(a, t2, f1);
+        }
+
+        if (sign1 < 0)
+            fmpz_neg(a, a);
+        if (sign2 < 0)
+            fmpz_neg(b, b);
+
+        fmpz_clear(t1);
+        fmpz_clear(t2);
     }
-    else if (!COEFF_IS_MPZ(*f))  /* only f is small */
-    {
-        mpz_t mf;
-        ulong tf = FLINT_ABS(*f);
-
-        mf->_mp_d = (mp_limb_t *) &tf;
-        mf->_mp_size  = fmpz_sgn(f);
-
-        mpz_gcdext(COEFF_TO_PTR(*d), COEFF_TO_PTR(*a), COEFF_TO_PTR(*b),
-                   mf, COEFF_TO_PTR(*g));
-    }
-    else if (!COEFF_IS_MPZ(*g))  /* only g is small */
-    {
-        mpz_t mg;
-        ulong tg = FLINT_ABS(*g);
-
-        mg->_mp_d = (mp_limb_t *) &tg;
-        mg->_mp_size  = fmpz_sgn(g);
-
-        mpz_gcdext(COEFF_TO_PTR(*d), COEFF_TO_PTR(*a), COEFF_TO_PTR(*b),
-                   COEFF_TO_PTR(*f), mg);
-    }
-    else /* both are big */
-    {
-        mpz_gcdext(COEFF_TO_PTR(*d), COEFF_TO_PTR(*a), COEFF_TO_PTR(*b),
-                   COEFF_TO_PTR(*f), COEFF_TO_PTR(*g));
-    }
-
-    _fmpz_demote_val(d);
-    _fmpz_demote_val(a);
-    _fmpz_demote_val(b);
 }
